feat: removed &nbsp; escapes in svg export (#482)
Previously, when the SVG export code thought it was outputting for an
html context it would escape spaces as &nbsp; When used via the
Javascript interface it would always do this. But this caused problems
when using the javascript interface to generate an SVG.

The purpose of escaping spaces is the ensure that whitespace is preserved
under html space collapsing rules. But the more ideal approach is to use
the css property white-space: pre to disable this.